cls_rbd: make get_parent return valid data when layering is disabled
This means clients can treat an error in their multi-object
transaction as a failure for all of them. This makes the client side
much simpler since it can call get_parent on images that don't support
layering, or do support it but don't have parents, and not need to
check the return value of every operation in the transaction.

Signed-off-by: Josh Durgin <josh.durgin@inktank.com>