librbd: return an error when removing a non-existent image
Try treating the image as new format if it's not in the old-style
directory, which is the last step in old-style removal. Then if the
image is not found in the new-style directory, -ENOENT will be
returned, preserving the semantics that existed prior to
cba4139d9a4d91d4cd2e41c01d440ebe176a339b.

Signed-off-by: Josh Durgin <josh.durgin@inktank.com>