mon: share global version with proposal, collect, commits, slurp, etc.
Share the global version through all of the paxos messages that share
data, and when slurping.  This ensures that committed values get the gv's
associated with them, and that proposed values, if they are later committed
by a new leader, will retain their original gv, and that new gv's will only
be allocated after that -- once paxos settles and things go writeable.

Signed-off-by: Sage Weil <sage@inktank.com>