rgw: fix rgw_dir_suggest_changes() to set header_changed
Apparently we weren't setting header_changed to true in the
case where we handled the CEPH_RGW_UPDATE case and cur_disk.exists
was false. In practice what this created is that in case where
object was created but the index complete call failed (or timed
out), calling rgw_dir_suggest_changes() fixed the entry, however,
we didn't account the new entry. This would lead to negative
stats on the bucket index.

Signed-off-by: Yehuda Sadeh <yehuda@inktank.com>