mds: Reset session stale timer at end of open
The mds gets an open session request and immediately sets the session
stale timer.  If logging the session and sending the reply takes longer
than the timer, the client sees a stale session message before getting
the open session reply.  This fix resets the session stale timer on
completion of the mds logging operation to reduce the likelihood of
the session getting marked stale before the client can receive the
open session reply and start sending renewcaps messages.

We reset the session timer for renewcaps and force_open_sessions right
before returning the SESSION_OPEN reply to the client.  This avoids stale
sessions in the cases where the mds is highly loaded and the session_open
reply back to the client is delayed.

Signed-off-by: Sam Lang <sam.lang@inktank.com>